Apply password and username obfuscation at runtime
It appears that the protocol employs some sort of username and password
obfuscation using XOR with a pattern contained within the the same message.
This change makes it explicit how that mechanism seems to operate and also
allows for guessing alternative usernames and non-numeric passwords.

Although this change opens more feature possibilities, it does not
contain any functional changes.